# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LedgerOS accounting system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: mine
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Submit expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
  /payroll:
    get:
      operationId: listPayroll
      tags: [payroll]
      summary: List payroll summaries
      responses:
        "200":
          description: Payroll summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollSummary"
    post:
      operationId: createPayroll
      tags: [payroll]
      summary: Add payroll summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollSummary"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payroll/{id}/transition:
    post:
      operationId: transitionPayroll
      tags: [payroll]
      summary: Submit or approve payroll summary
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionInput"
      responses:
        "200":
          description: Payroll summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollSummary"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reports/{type}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a financial report
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TransitionInput:
      type: object
      required: [action]
      properties:
        action: { type: string, minLength: 1 }
        note: { type: string }
        matchBankTransactionId: { type: ["integer", "null"] }
...
    Vendor:
      type: object
      required: [id, name, balance]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        isSensitive: { type: boolean }
        notes: { type: ["string", "null"] }
        balance: { type: number }
        isTestData: { type: boolean }
    VendorInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        category: { type: string }
        isSensitive: { type: boolean }
        notes: { type: string }
    VendorUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        category: { type: string }
        isSensitive: { type: boolean }
        notes: { type: string }
...
    Expense:
      type: object
      required: [id, description, amount, status, expenseDate]
      properties:
        id: { type: integer }
        description: { type: string }
        vendorName: { type: ["string", "null"] }
        amount: { type: number }
        category: { type: ["string", "null"] }
        clientProject: { type: ["string", "null"] }
        sourceApp: { type: ["string", "null"] }
        status: { type: string }
        expenseDate: { type: string }
        hasReceipt: { type: boolean }
        reimbursable: { type: boolean }
        submittedById: { type: ["integer", "null"] }
        submittedByName: { type: ["string", "null"] }
        isTestData: { type: boolean }
        createdAt: { type: string }
    ExpenseInput:
      type: object
      required: [description, amount, expenseDate]
      properties:
        description: { type: string, minLength: 1 }
        vendorName: { type: string }
        amount: { type: number }
        category: { type: string }
        clientProject: { type: string }
        sourceApp: { type: string }
        expenseDate: { type: string }
        hasReceipt: { type: boolean }
...
    PayrollSummary:
      type: object
      required: [id, periodStart, periodEnd, payDate, grossPay, status]
      properties:
        id: { type: integer }
        periodStart: { type: string }
        periodEnd: { type: string }
        payDate: { type: string }
        grossPay: { type: number }
        taxes: { type: number }
        netPay: { type: number }
        employerTaxes: { type: number }
        employeeCount: { type: ["integer", "null"] }
        source: { type: ["string", "null"] }
        status: { type: string }
        flagged: { type: boolean }
        flagReason: { type: ["string", "null"] }
        isTestData: { type: boolean }
    PayrollInput:
      type: object
      required: [periodStart, periodEnd, payDate, grossPay]
      properties:
        periodStart: { type: string }
        periodEnd: { type: string }
        payDate: { type: string }
        grossPay: { type: number }
        taxes: { type: number }
        netPay: { type: number }
        employerTaxes: { type: number }
        employeeCount: { type: integer }
        source: { type: string }
...
    ReportRow:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
        group: { type: ["string", "null"] }
        detail: { type: ["string", "null"] }
    Report:
      type: object
      required: [type, title, rows]
      properties:
        type: { type: string }
        title: { type: string }
        generatedAt: { type: string }
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ReportRow"
        totals:
          type: array
          items:
            $ref: "#/components/schemas/ReportRow"
...